Fix ARM64 CFFI struct-by-value ABI bug; widen rounding/version enums to uint64_t
CFFI ABI mode (ffi.dlopen) on ARM64 mis-marshals structs whose sub-64-bit
fields pack into a single register. fdl_round_strategy_t (2 x uint32_t = 8B)
had both fields packed into x0, causing them to be swapped when passed as a
parameter. This broke all four Python call sites that pass rounding by value:
fdl_doc_add_canvas_template, fdl_compute_framing_from_intent,
fdl_framing_decision_populate_from_intent, and fdl_geometry_round.

Fix: widen fdl_rounding_mode_t, fdl_rounding_even_t, and fdl_abi_version_t
fields from uint32_t to uint64_t. Each field now gets its own GPR (x0, x1),
matching fdl_dimensions_i64_t which is confirmed working. This replaces a
prior scalar-accessor workaround that only covered 1 of 4 affected call sites.

- Add 20 struct-by-value ABI regression tests (all struct types, all directions)
- Add 5 viewer rounding round-trip tests
- Document the uint64_t rationale in fdl_core.h enum section header

+++ native/bindings/python/fdl_ffi/fdl_core_decl.h
@@ -18,9 +18,9 @@ typedef uint32_t fdl_custom_attr_type_t;
 #define FDL_CUSTOM_ATTR_TYPE_OTHER 8
 
 typedef struct fdl_abi_version_t {
-    uint32_t major;
-    uint32_t minor;
-    uint32_t patch;
+    uint64_t major;
+    uint64_t minor;
+    uint64_t patch;
 } fdl_abi_version_t;
 
 fdl_abi_version_t fdl_abi_version(void);
@@ -47,12 +47,12 @@ typedef struct fdl_rect_t {
     double height;
 } fdl_rect_t;
 
-typedef uint32_t fdl_rounding_mode_t;
+typedef uint64_t fdl_rounding_mode_t;
 #define FDL_ROUNDING_MODE_UP 0
 #define FDL_ROUNDING_MODE_DOWN 1
 #define FDL_ROUNDING_MODE_ROUND 2
 
-typedef uint32_t fdl_rounding_even_t;
+typedef uint64_t fdl_rounding_even_t;
 #define FDL_ROUNDING_EVEN_WHOLE 0
 #define FDL_ROUNDING_EVEN_EVEN 1
